% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility.R
\docType{class}
\name{Facility-class}
\alias{Facility-class}
\title{Facility Class}
\description{
A class to represent a facility with attributes like name, sector, location, and MACC data.
}
\section{Slots}{

\describe{
\item{\code{Name}}{A string representing the name of the facility.}

\item{\code{Sector}}{A string representing the sector (e.g., Steel, Cement).}

\item{\code{Location}}{A string representing the city where the facility is located.}

\item{\code{Location_Lat_Long}}{A data frame containing Latitude and Longitude of the location.}

\item{\code{Size}}{A numeric value representing the size of the facility.}

\item{\code{Baseline_CO2_Intensity}}{A numeric value representing baseline CO2 intensity (tCO2/tOutput).}

\item{\code{Target_Intensity}}{A numeric value representing target CO2 intensity (tCO2/tOutput).}

\item{\code{MACC}}{A data frame with two columns: CO2 price and Abatement Potential.}
}}

